// src/app/core/models/brand.model.ts
export interface Brand {
    id: string;
    owner: null | string;
    industry: string;
    name: string;
    description: null | string;
    countryOrigin: null | string;
    website: null | string;
    isDeleted: null;
    isActive: boolean;
    createdBy: null;
    createdDate: null;
    lastModifiedBy: null;
    lastModifiedDate: null;
    companyId: string;
}